+++ CHANGELOG.md
@@ -7,8 +7,8 @@ All Notable changes to `Csv` will be documented in this file
 ### Added
 
 - `League\Csv\TabularDataWriter` interface to represent how to write to a tabular data document.
-- `TabularDataReader::fetchFirst` to replace `TabularDataReader::fetchOne`
-- `TabularDataReader::fetchNth` to replace `TabularDataReader::fetchOne`
+- `TabularDataReader::first` to replace `TabularDataReader::fetchOne`
+- `TabularDataReader::nth` to replace `TabularDataReader::fetchOne`
 
 ### Deprecated